flaskapp/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==2.0.1
python-dotenv==0.15.0
PyYAML==5.3.1
SQLAlchemy==1.3.20
Werkzeug==1.0.1
